Refactor entrypoints into seperate files and update entrypoint names
